move factorial* to scipy.special
