+++ scipy/misc/common.py
@@ -5,14 +5,11 @@ Functions which are common and require SciPy Base and Level 1 SciPy
 
 from __future__ import division, print_function, absolute_import
 
-from scipy.lib.six import xrange
-
 from numpy import exp, log, asarray, arange, newaxis, hstack, product, array, \
-                  where, zeros, extract, place, pi, sqrt, eye, poly1d, dot, \
-                  r_, rollaxis, sum, fromstring, ndarray
+                  zeros, eye, poly1d, r_, rollaxis, sum, fromstring
 
-__all__ = ['logsumexp', 'factorial','factorial2','factorialk',
-           'central_diff_weights', 'derivative', 'pade', 'lena', 'ascent', 'face']
+__all__ = ['logsumexp', 'central_diff_weights', 'derivative', 'pade', 'lena',
+           'ascent', 'face']
 
 # XXX: the factorial functions could move to scipy.special, and the others
 # to numpy perhaps?
@@ -88,159 +85,6 @@ def logsumexp(a, axis=None, b=None):
         out = log(sum(exp(a - a_max), axis=0))
     out += a_max
     return out
-
-
-def factorial(n,exact=False):
-    """
-    The factorial function, n! = special.gamma(n+1).
-
-    If exact is 0, then floating point precision is used, otherwise
-    exact long integer is computed.
-
-    - Array argument accepted only for exact=False case.
-    - If n<0, the return value is 0.
-
-    Parameters
-    ----------
-    n : int or array_like of ints
-        Calculate ``n!``.  Arrays are only supported with `exact` set
-        to False.  If ``n < 0``, the return value is 0.
-    exact : bool, optional
-        The result can be approximated rapidly using the gamma-formula
-        above.  If `exact` is set to True, calculate the
-        answer exactly using integer arithmetic. Default is False.
-
-    Returns
-    -------
-    nf : float or int
-        Factorial of `n`, as an integer or a float depending on `exact`.
-
-    Examples
-    --------
-    >>> arr = np.array([3,4,5])
-    >>> sc.factorial(arr, exact=False)
-    array([   6.,   24.,  120.])
-    >>> sc.factorial(5, exact=True)
-    120L
-
-    """
-    if exact:
-        if n < 0:
-            return 0
-        val = 1
-        for k in xrange(1,n+1):
-            val *= k
-        return val
-    else:
-        from scipy import special
-        n = asarray(n)
-        sv = special.errprint(0)
-        vals = special.gamma(n+1)
-        sv = special.errprint(sv)
-        return where(n >= 0,vals,0)
-
-
-def factorial2(n, exact=False):
-    """
-    Double factorial.
-
-    This is the factorial with every second value skipped, i.e.,
-    ``7!! = 7 * 5 * 3 * 1``.  It can be approximated numerically as::
-
-      n!! = special.gamma(n/2+1)*2**((m+1)/2)/sqrt(pi)  n odd
-          = 2**(n/2) * (n/2)!                           n even
-
-    Parameters
-    ----------
-    n : int or array_like
-        Calculate ``n!!``.  Arrays are only supported with `exact` set
-        to False.  If ``n < 0``, the return value is 0.
-    exact : bool, optional
-        The result can be approximated rapidly using the gamma-formula
-        above (default).  If `exact` is set to True, calculate the
-        answer exactly using integer arithmetic.
-
-    Returns
-    -------
-    nff : float or int
-        Double factorial of `n`, as an int or a float depending on
-        `exact`.
-
-    Examples
-    --------
-    >>> factorial2(7, exact=False)
-    array(105.00000000000001)
-    >>> factorial2(7, exact=True)
-    105L
-
-    """
-    if exact:
-        if n < -1:
-            return 0
-        if n <= 0:
-            return 1
-        val = 1
-        for k in xrange(n,0,-2):
-            val *= k
-        return val
-    else:
-        from scipy import special
-        n = asarray(n)
-        vals = zeros(n.shape,'d')
-        cond1 = (n % 2) & (n >= -1)
-        cond2 = (1-(n % 2)) & (n >= -1)
-        oddn = extract(cond1,n)
-        evenn = extract(cond2,n)
-        nd2o = oddn / 2.0
-        nd2e = evenn / 2.0
-        place(vals,cond1,special.gamma(nd2o+1)/sqrt(pi)*pow(2.0,nd2o+0.5))
-        place(vals,cond2,special.gamma(nd2e+1) * pow(2.0,nd2e))
-        return vals
-
-
-def factorialk(n,k,exact=True):
-    """
-    n(!!...!)  = multifactorial of order k
-    k times
-
-    Parameters
-    ----------
-    n : int, array_like
-        Calculate multifactorial. Arrays are only supported with exact
-        set to False. If `n` < 0, the return value is 0.
-    exact : bool, optional
-        If exact is set to True, calculate the answer exactly using
-        integer arithmetic.
-
-    Returns
-    -------
-    val : int
-        Multi factorial of `n`.
-
-    Raises
-    ------
-    NotImplementedError
-        Raises when exact is False
-
-    Examples
-    --------
-    >>> sc.factorialk(5, 1, exact=True)
-    120L
-    >>> sc.factorialk(5, 3, exact=True)
-    10L
-
-    """
-    if exact:
-        if n < 1-k:
-            return 0
-        if n <= 0:
-            return 1
-        val = 1
-        for j in xrange(n,0,-k):
-            val = val*j
-        return val
-    else:
-        raise NotImplementedError
 
 
 def central_diff_weights(Np, ndiv=1):
